Update capistrano to 3.8.0
